# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BookHive — Digital Books & Audio Marketplace",
    "description": "BookHive is a marketplace for buying, selling and sharing digital books (PDF) and audiobooks. This API covers user authentication, catalog management, file uploads (audio/pdf), purchases, subscriptions and reviews.",
    "termsOfService": "https://bookhive.example.com/terms",
    "contact": {
      "name": "BookHive API Team",
      "email": "api@bookhive.example.com",
      "url": "https://bookhive.example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "2025-11-13"
  },
  "servers": [
    { "url": "http://localhost:8000/v1", "description": "Local dev" },
    { "url": "https://api.bookhive.com/v1", "description": "Production" }
  ],
...
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "Register new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "User created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Login and receive JWT token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } }
            }
          },
          "401": { "description": "Invalid credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "apiKey": { "type": "apiKey", "in": "header", "name": "x-api-key" }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["email","password","displayName"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "alice@example.com" },
          "password": { "type": "string", "minLength": 8 },
          "displayName": { "type": "string", "example": "Alice Reader" },
          "bio": { "type": "string" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email","password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "tokenType": { "type": "string", "example": "Bearer" },
          "expiresIn": { "type": "integer", "example": 3600 }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "usr_01F..." },
          "email": { "type": "string", "format": "email" },
          "displayName": { "type": "string" },
          "bio": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "code": { "type": "integer" },
          "message": { "type": "string" },
          "details": { "type": "object" }
        }
      }
    }
  },
  "security": [{ "bearerAuth": [] }]
}